requirements.txt
requests
python-dotenv
PyMuPDF      # For PDF reading
llama-index-llms-ollama


